+++ src/main/java/chess/dao/BoardDao.java
@@ -1,6 +1,6 @@
 package chess.dao;
 
-import chess.entity.BoardEntity;
+import chess.dao.entity.BoardEntity;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
